Функции создания БД вынесены из файла тестирования в отдельный файл
