+++ tests/account/download/report__test.py
@@ -2,67 +2,12 @@ from datetime import datetime
 
 import pytest
 
-from django.contrib.auth.models import User
-
 from account.report import CityReport, RegionReport, AreaReport
-from city.models import City, VisitedCity
-from region.models import Region, Area
-
-
-def create_user(django_user_model):
-    return django_user_model.objects.create_user(id=1, username='username', password='password')
-
-
-def create_area(num: int) -> list[Area]:
-    return [
-        Area.objects.create(id=1, title=f'Округ {i}') for i in range(1, num + 1)
-    ]
-
-
-def create_region(num: int, area: Area) -> list[Region]:
-    return [
-        Region.objects.create(
-            id=1,
-            area=area,
-            title=f'Регион {i}',
-            type='область',
-            iso3166=f'RU-RU{i}'
-        ) for i in range(1, num + 1)
-    ]
-
-
-def create_city(num: int, region: Region) -> list[City]:
-    return [
-        City.objects.create(
-            id=i,
-            title=f'Город {i}',
-            region=region,
-            coordinate_width=1,
-            coordinate_longitude=1
-        ) for i in range(1, num + 1)
-    ]
-
-
-def create_visited_city(
-        region: Region,
-        user: User,
-        city: City,
-        date_of_visit: datetime,
-        has_magnet: bool,
-        rating: int
-) -> VisitedCity:
-    return VisitedCity.objects.create(
-        user=user,
-        region=region,
-        city=city,
-        date_of_visit=date_of_visit,
-        has_magnet=has_magnet,
-        rating=rating
-    )
+from tests.account.download.create_db import create_user, create_area, create_region, create_city, create_visited_city
 
 
 @pytest.fixture
-def setup_db(client, django_user_model):
+def setup_db(django_user_model):
     user = create_user(django_user_model)
     area = create_area(1)
     region = create_region(1, area[0])
